### 2025-04-23 #8100 @editorialbot
[PRE REVIEW]: CATS: The Climate Aware Task Scheduler
**Submitting author:** @andreww (<a href="http://orcid.org/0000-0003-3121-3255">Andrew Walker</a>)
**Repository:** https://github.com/GreenScheduler/cats
**Branch with paper.md** (empty if default branch): paper
**Version:** 1.0.0
