### 2025-04-23 #8100 @kyleniemeyer
Hi @andreww, before we proceed with finding an editor and reviewers, the JOSS editorial board is going to take a closer look at your submission to make sure it falls in our scope as [research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software). We should get back to you within a week or two on this.
