### 2025-05-01 #8100 @christopheredsall
> Hi there [@christopheredsall](https://github.com/christopheredsall)!
>
> This paper has been submitted to JOSS, the Journal of Open Source Software. It is a job scheduler designed to maximise the use of green energy in compute jobs.
>
> Would you be willing to review it? Please indicate by replying to this thread.
>
> You can find information about reviewing for JOSS [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

Hi @jatkinson1000 ,

Sure I can give that a go.

Of note, I am at the same institution as one of the authors (Loïc Lannelongue), however in completely different Institutions/departments and the university has 12 000 staff! We have never had

> collaborators in funded research or work that is published

and I think that, per the https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy I could

> make an impartial scientific judgment or evaluation.

If that is still OK by you.
