+++ docs/paper/paper.bib
@@ -22,7 +22,7 @@ journal = {Advanced Science},
 volume = {8},
 number = {12},
 pages = {2100707},
-doi = {https://doi.org/10.1002/advs.202100707},
+doi = {10.1002/advs.202100707},
 year = {2021}
 }
 
