### 2025-05-02 fc81cfd
First pass at "approach and implementation" docs
This is lifted from the manuscript and converted into
restructured text (if I have that correct).
