+++ docs/source/approach.rst
@@ -1,0 +1,2 @@
+.. _approach:
+
