+++ docs/source/index.rst
@@ -14,6 +14,7 @@ cluster jobs to minimize predicted carbon intensity of running the process.
    introduction.rst
    installation.rst
    quickstart.rst
+   approach.rst
    use-with-schedulers.rst
    cli-reference.rst
    api-reference.rst
