### 2025-05-02 ca6967a
First pass at trimming the paper
