+++ docs/paper/paper.md
@@ -77,19 +77,13 @@ There are several approaches that have been proposed and/or implemented to reduc
 
 # Approach and functionality
 
-At its core CATS is an open-source (MIT licence) Python package (tested with Python 3.9-3.12) that combines data on the forecast carbon intensity of the local electricity supply with information about a proposed job's duration to assess the best start time of the computation within the validity interval of the forecast. Users typically interact with CATS via a command-line interface targeting the UNIX Shell (CATS is tested on Linux and MacOS) and the best start time can be provided in an informative format (that the user can then use with their infrastructure) or in a way that can be passed on to job schedulers to set the calculation start time. CATS is available via the Python Package Index (PyPI) and can be installed along with its handful of dependencies into a Python environment with pip. Development takes place on GitHub (https://github.com/GreenScheduler/cats) and documentation is available (https://greenscheduler.github.io/cats/).
+At its core CATS is an open-source (MIT licence) Python package (tested with Python 3.9-3.12) that combines data on the forecast carbon intensity of the local electricity supply with information about a proposed job's duration to assess the best start time of the computation within the validity interval of the forecast. Users typically interact with CATS via a command-line interface targeting the UNIX Shell (CATS is tested on Linux and MacOS) and the best start time can be provided in an informative format (that the user can then use with their infrastructure) or in a way that can be passed on to job schedulers to set the calculation start time. CATS is available via the Python Package Index (PyPI) and can be installed along with its handful of dependencies into a Python environment with pip. Development takes place on GitHub (https://github.com/GreenScheduler/cats) and documentation, including a fuller description of the approach, is available (https://greenscheduler.github.io/cats/).
 
-At a minimum, the user must provide CATS with the duration of the proposed computation on the command line and CATS also requires information about the location where the computation is to happen. This can be provided on the command line, via a configuration file, or from geolocation of the IP address and is sufficient information for CATS to access a prediction of the carbon intensity of the relevant power distribution network which is used to compute the start time that minimises the carbon intensity over the duration of the computation.
-
-At the moment, CATS is available for the UK through the National Grid's API which provides postcode-specific 48 hour forecasts broken down into 30 minute periods. This granular data contains regional distribution networks making use of a parameterised model of the power distribution system in Great Britain, weather forecasts and historical generation data. A brief overview of the forecast methodology can be found in [@Bruce21a; @Bruce21b] and the forecast, in the form of estimated carbon intensity at the end of each half hour period, is available via a web API (https://carbonintensity.org.uk/). CATS caches requests for this data to avoid repeated requests within the thirty minute time frame of a single forecast. We have designed CATS in a modular way to enable future integration of other countries' APIs in a straightforward manner. However, to date, we are unaware of any publicly accessible forecasts of carbon intensity for other regional, national or transnational electricity distribution networks.
-
-With the carbon intensity forecast and duration of the proposed computation in hand, the next task is to locate the start time (within the valid forecast period) that minimises the integrated carbon intensity over a sliding window equal to the duration. All else being equal, this minimising start time is the time when the carbon cost of the proposed computation will be least. The difference in the integrated carbon footprint between an optimal start time and starting the calculation immediately is a measure of the potential benefits of timeshifting the computation. An illustration of this calculation is provided in Figure 1. Once the carbon intensity minimisation has been completed, CATS can optionally submit the computation to a queueing system or make a more detailed report on the climate impact of the proposed computation.
-
[…]
