### 2025-05-02 #8100 @andreww
Thanks @jatkinson1000!

I've fixed the broken DOI and checked that the other three don't have one. I've also given trimming the paper a first go - mostly by trimming the details on how CATS works (and moving that to a new page of our documentation). We've got our regular monthly developers meeting this afternoon so I'll use that to see if we can sugest reviewers and look to see if we can shorten the manuscript further.
