### 2025-05-02 #8100 @jatkinson1000
Perfect, thanks @andreww

Yes - I should have said please do keep information that gets cut from the paper available somewhere in the wider docs as I'm sure it is useful.

Also note that you can place the following command in a comment on this thread to regenerate the paper pdf with your updates `@editorialbot generate pdf`.
