### 2025-05-02 d7dd0b2
Actually include the text in the new approach segment of the documentation
Need to deal with the restructured text formatting!
