+++ docs/source/approach.rst
@@ -1,2 +1,88 @@
 .. _approach:
 
+Approach and Implementation
+===========================
+
+There are several approaches that have been proposed and/or implemented to reduce the carbon footprint 
+of computing. These include interventions in procurement to maximise the useful life of computer 
+hardware and thus reduce the embodied carbon cost; exercises in code modification to reduce the 
+carbon cost of a particular calculation, mostly by shortening the runtime or reducing resource needs; 
+modifying default settings to reduce carbon use [e.g. the clock frequency on ARCHER2, the UK national 
+supercomputer, is now lower by default, @Jackson23]; and attempts to improve data centre efficiency. 
+However, many of these can be difficult for individual researchers to implement. One 
+easier-to-implement approach is to make use of lower carbon intensive electricity to power the 
+computation. This can be achieved either by relocating the computer to a part of the world where 
+electricity is generated using renewable means [e.g. the University of York in the UK have located 
+the latest institutional computing resource in Sweden @Norsecode], or by time shifting the computation 
+such that it runs when the power supplied by the local electricity grid is dominated by renewable 
+generation such as on windy or sunny days rather than periods where legacy fossil fuel generation 
+dominates. The potential impact of minimally-invasive time-shifting approaches has been shown to 
+result in significantly reduced carbon footprints [27% in one AI benchmark @Dodge2022]. In a country 
+like the UK, these reductions can reach over 60% by shifting workload by a day or two [@ElectricityMaps]. 
+CATS helps researchers timeshift their own computation such that it is scheduled when the forecast 
+carbon intensity of the power grid is minimised.
+
+At its core CATS is an open-source (MIT licence) Python package (tested with Python 3.9-3.12) that 
+combines data on the forecast carbon intensity of the local electricity supply with information about 
+a proposed job's duration to assess the best start time of the computation within the validity interval 
+of the forecast. Users typically interact with CATS via a command-line interface targeting the UNIX 
+Shell (CATS is tested on Linux and MacOS) and the best start time can be provided in an informative 
+format (that the user can then use with their infrastructure) or in a way that can be passed on to job 
+schedulers to set the calculation start time. CATS is available via the Python Package Index (PyPI) 
+and can be installed along with its handful of dependencies into a Python environment with pip. 
+Development takes place on GitHub (https://github.com/GreenScheduler/cats) and documentation is
+ available (https://greenscheduler.github.io/cats/).
+
+At a minimum, the user must provide CATS with the duration of the proposed computation on the command 
+line and CATS also requires information about the location where the computation is to happen. This 
+can be provided on the command line, via a configuration file, or from geolocation of the IP address 
+and is sufficient information for CATS to access a prediction of the carbon intensity of the relevant 
+power distribution network which is used to compute the start time that minimises the carbon intensity 
+over the duration of the computation.
+
+At the moment, CATS is available for the UK through the National Grid's API which provides 
+postcode-specific 48 hour forecasts broken down into 30 minute periods. This granular data contains 
+regional distribution networks making use of a parameterised model of the power distribution system 
+in Great Britain, weather forecasts and historical generation data. A brief overview of the forecast 
+methodology can be found in [@Bruce21a; @Bruce21b] and the forecast, in the form of estimated carbon 
+intensity at the end of each half hour period, is available via a web API 
+(https://carbonintensity.org.uk/). CATS caches requests for this data to avoid repeated requests 
+within the thirty minute time frame of a single forecast. We have designed CATS in a modular way to 
+enable future integration of other countries' APIs in a straightforward manner. However, to date, we 
+are unaware of any publicly accessible forecasts of carbon intensity for other regional, national or 
+transnational electricity distribution networks.
+
+With the carbon intensity forecast and duration of the proposed computation in hand, the next task 
[…]
