### 2025-05-02 2e9f5fa
Further shorten paper
Reduce the statment of need to avoid unbalancing things.
