### 2025-05-02 #8100 @andreww
We've trimmed 635 words from the paper so ignoring the affiliations and acknowledgments we're down below 1000 words. A couple of the co-authors are going to take a look over the weekend but I think we're where we need to be.

In terms of reviewers, members of the [Green Research Software Engineering SIG](https://socrse.github.io/green-sig/) should be well placed. jwallwork23 is on the SIG steering committee and is a JOSS reviewer with knowledge of python, but is based in Cambridge. Of other reviewers on the system, kanishkan91 mentions carbon accounting and python and ali-ramadhan mentions climate science, scientific computing and python. @Llannelongue is our co-author with the best view on the literature if you need additional suggestions from outside the list on the JOSS system.
