### 2025-05-02 #8100 @jatkinson1000
Hi there @KirstyPringle, @kellyrowland !

This paper has been submitted to JOSS, the Journal of Open Source Software.
It is a job scheduler designed to maximise the use of green energy in compute jobs.

Would you be willing to review it?
Please indicate by replying to this thread.

You can find information about reviewing for JOSS [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
