### 2025-05-12 d6b8fd5
capturing and printing the output of sbatch to let the user know the job number
