### 2025-05-12 33f499c
preventing b' being shown at the start of the slurm output
