### 2025-05-13 #8100 @jatkinson1000
Hi there @torressa and @nightlark!

This paper has been submitted to JOSS, the Journal of Open Source Software.
It is a job scheduler designed to maximise the use of green energy in compute jobs based on real-time electricity grid data.

Would you be willing to review it?
Please indicate by replying to this thread.

You can find information about reviewing for JOSS [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
