### 2025-05-19 #8100 @LiamPattinson
Hi @jatkinson1000, I would be happy to review, although I expect that I'll be very busy for the next two weeks. I should have time starting in June. It would also be my first JOSS review, so I may need some advice and assistance from time to time.
