### 2025-05-19 #8100 @jackleland
Hi @jatkinson1000, I would be keen to review but I think I have a conflict of interest: one of the authors (Abhishek Dasgupta) is a colleague of mine on the Oxford Research Software Engineering team. I'm not sure where the line is but I think it would be difficult for my assessment of the work to be completely impartial - happy to discuss though of course
