### 2025-05-19 #8100 @jatkinson1000
No problem - it is a checkbased process and there is nice guidance on the JOSS site.
We usually aim for a 4-week turnaround, so that should be fine.
Thank you for your service!
