### 2025-05-19 #8100 @jatkinson1000
Alas @jackleland that would be a CoI under the JOSS policy - thanks for letting me know and hopefully we'll get you to review something else soon!
