### 2025-05-19 #8100 @jatkinson1000
Hi @dokempf,

This paper has been submitted to JOSS, the Journal of Open Source Software.
It is a job scheduler, written in Python, designed to maximise the use of green energy in compute jobs based on real-time electricity grid data.

Would you be willing to review it?
Please indicate by replying to this thread.

You can find information about reviewing for JOSS [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
