### 2025-05-19 #8100 @jatkinson1000
Hi @aturner-epcc that would be fine.
We usually aim for a 4-week review period (with open discussion between reviewers and authors) which would take us to 16th June from today, and we could extend this given you have let us know ahead of time.
