### 2025-06-03 #8251 @LiamPattinson
I've finished my first pass over the checklist and identified a few issues -- some that I think require corrections, others where I just need a clarification. I'm testing the functionality now, and will know if it worked in a couple of hours after my scheduled job goes through. I'll raise some issues on the repository tomorrow.
