### 2025-06-04 GreenScheduler/cats#143 @LiamPattinson
JOSS Review: Community guidelines
Hi there, I'm currently reviewing the JOSS paper for CATS (openjournals/joss-reviews#8251), and overall the repository is looking good! I'll be raising a couple of issues besides this one to make comments and suggest changes, but I don't think any major changes will be necessary.

One element of our reviewers checklist is the following:

> - [ ] Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

I see that your `README.md` contains the following text:

```
We welcome contributions from the community! If you find a bug or
have an idea for a new feature, please open an issue on our GitHub
repository or submit a pull request.
```

I think it would be helpful to expand this section to give more information for users seeking general support with the software. I'd also recommend adding a `CONTRIBUTING.md` file to the repository, an example of which can be generated at [https://contributing.md/example/](https://contributing.md/example/).
