### 2025-06-04 GreenScheduler/cats#144 @LiamPattinson
JOSS Review: Recommend new PyPI release
I ran into an issue while reviewing CATS for the JOSS review (https://github.com/openjournals/joss-reviews/issues/8251) regarding this point on our checklist:

> - [ ] Functionality: Have the functional claims of the software been confirmed?

I believe CATS does meet this requirement when installed from source, but the version obtained when installing via `pip install climate-aware-task-scheduler` is lacking some key features and was quite buggy (when running with low durations of `-d < 20`, I found it would sometimes fail with an index out of bounds error and no further clarifying text). As the last PyPI release was in July 2024, I'd recommend making a new release soon.
