### 2025-06-06 GreenScheduler/cats#143 @aturner-epcc
As another JOSS reviewer, I completely agree with this comment.
