### 2025-06-06 GreenScheduler/cats#144 @aturner-epcc
Confirm that I also see the same issue with the PyPI release:

```
❯ cats --duration 10 --location "EH8"
WARNING:root:config file not found
WARNING:root:Unspecified carbon intensity forecast service, using carbonintensity.org.uk
Traceback (most recent call last):
  File "/Users/aturner/anaconda3/bin/cats", line 8, in <module>
    sys.exit(main())
             ^^^^^^
  File "/Users/aturner/anaconda3/lib/python3.11/site-packages/cats/__init__.py", line 289, in main
    now_avg, best_avg = get_avg_estimates(CI_forecast, duration=duration)
                        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/aturner/anaconda3/lib/python3.11/site-packages/cats/optimise_starttime.py", line 26, in get_avg_estimates
    return wf[0], min(wf)
                  ^^^^^^^
  File "/Users/aturner/anaconda3/lib/python3.11/site-packages/cats/forecast.py", line 147, in __iter__
    yield self.__getitem__(index)
          ^^^^^^^^^^^^^^^^^^^^^^^
  File "/Users/aturner/anaconda3/lib/python3.11/site-packages/cats/forecast.py", line 97, in __getitem__
    self.data[index + 1],
    ~~~~~~~~~^^^^^^^^^^^
IndexError: list index out of range
```
