### 2025-06-06 #8251 @aturner-epcc
@andreww One thing that seems to be missing from the paper is any comparison to other software that enables temporal shifting to reduce emissions.
