### 2025-06-06 GreenScheduler/cats#143 @andreww
Thanks. We've had a [contributor code of conduct](https://github.com/GreenScheduler/cats/blob/main/CODE_OF_CONDUCT.md) in the repository from day one but that certainly needs to make it into the documentation too. I also agree that we need something thats more of a guide for contributors and I'll get that added. Thanks so much for the link to contributing.md - looks super helpful.
