### 2025-06-06 GreenScheduler/cats#144 @andreww
Thanks - @abhidg is currently setting up a workflow to manage the release process to PyPi (issue #141). We'll get that done and update the version on PyPi.
