### 2025-06-06 11415c8
Allow cats_config.yml and cats_config.yaml
This change allows us to read configuration data from local files
cats_config.yml, cats_config.yaml or config.yml. The first file
found (in that order) is read and any others are ignored.

I've also updated the tests to check that this works which involved
fun with fixtures. Having thought about it a bit more this could
probably be cleaned up with a parameterisation but I'm not sure
about extracting the parameter for the test in the fixture ...
something for another day.
