+++ pyproject.toml
@@ -38,6 +38,9 @@
   [tool.setuptools.dynamic]
     version = {attr = "cats.__version__"}
 
+  [tool.pytest.ini_options]
+    markers = ["fixt_data: data for fixtures"]
+
   [project.optional-dependencies]
     test = ["pytest", "numpy>=1.5.0", "pytest-subprocess==1.5.0", "pytest-cov"]
     types = ["mypy", "types-PyYAML", "types-redis", "types-requests", "types-ujson"]
