### 2025-06-16 e310c9b
Add sphinx conf to RTD cof
Read the docs needs to know how to find sphinx config
