+++ .readthedocs.yaml
@@ -4,6 +4,10 @@
 # Required
 version: 2
 
+sphinx:
+  # Path to your Sphinx configuration file.
+  configuration: docs/source/conf.py
+
 # Set the OS, Python version, and other tools you might need
 build:
   os: ubuntu-24.04
