### 2025-06-16 4c4a0fd
Lets install cats before we buld RTD
