+++ .readthedocs.yaml
@@ -4,6 +4,13 @@
 # Required
 version: 2
 
+python:
+  install:
+    - method: pip
+      path: .
+      extra_requirements:
+        - docs
+
 sphinx:
   # Path to your Sphinx configuration file.
   configuration: docs/source/conf.py
