### 2025-06-17 d8e76f2
Create draft CONTRIBUTING.md
We may actually want to move this to read the docs, but put it here
for now (and to check the PR documentation generator works).

This is generated using contributing.md and I've stripped out the following
parts (which we may want to put back if we have content):

```
  ### Your First Code Contribution
  <!-- TODO
  include Setup of env, IDE and typical getting started instructions?

  -->

  ### Improving The Documentation
  <!-- TODO
  Updating, improving and correcting the documentation

  -->

  ## Styleguides
  ### Commit Messages
  <!-- TODO

  -->

  ## Join
[…]
