+++ cats/__init__.py
@@ -16,7 +16,7 @@ from .configure import get_runtime_config
 from .constants import CATS_ASCII_BANNER_COLOUR, CATS_ASCII_BANNER_NO_COLOUR
 from .forecast import CarbonIntensityAverageEstimate, WindowedForecast
 
-__version__ = "1.0.0"
+__version__ = "1.0.2"
 
 # To add a scheduler, add a date format here
 # and create a scheduler_<new>(...) function
