### 2025-06-17 #8251 @LiamPattinson
Hi @jatkinson1000, I was thinking of raising an issue to ask about the contributor list, as one or two people may be missing, and I'll have to double check to see if the references follow the citation syntax properly, but that's all from me.
