### 2025-06-17 #8251 @andreww
> Hi [@jatkinson1000](https://github.com/jatkinson1000), I was thinking of raising an issue to ask about the contributor list, as one or two people may be missing

Before submission I emailed everybody who were either involved in the CW23 hack day or who had contributed to the repository if they wanted to be included in the paper and asked for their ORCID (I also sent a copy of the manuscript). One declined as they felt their contribution was too small, and another never replied.
