### 2025-06-24 f233e08
Fix spacing issue in help output
Co-authored-by: Sadie L. Bartholomew <sadie.bartholomew@ncas.ac.uk>
