### 2025-06-24 9a0f85d
Add material on other timeshifting approaches.
(and minor cleanup).
