+++ docs/paper/paper.bib
@@ -85,3 +85,16 @@ series = {SC-W '23}
 
  @misc{MacBookPro, author={{Apple Inc}}, year={2024}, url={https://www.apple.com/environment/pdf/products/notebooks/MacBook_Pro_16-inch_PER_Oct2024.pdf}, language={en} }
 
+ @misc{Norlander, author={Nordlander, T. F. and Lybecker, A.}, year={2023}, url={https://devblogs.microsoft.com/ise/saving-co2-using-location-and-time-shifting-in-azure/}, title={{Saving CO2 using location and time shifting in Azure
+}}
+
+ @article{Wiesner2021, author={Wiesner, P. and Behnke, I. and Scheinert, D. and Gontarska, K. and Thamsen, L.}, 
+          year={2021}, title={{Let's wait awhile: how temporal workload shifting can reduce carbon emissions in the cloud}}, 
+          doi={10.1145/3464298.349339}, journal={{Middleware '21: Proceedings of the 22nd International Middleware Conference}}, 
+          pages={{260-272}}}
+
+ @article{Jagannadharao2025, author={Jagannadharao, A. and Beckage, N. and Nafus, D. and Chamberlin, S.},
+          year={2025}, title={Timeshifting strategies for carbon-efficient long-running large language model training},
+          doi={10.1007/s11334-023-00546-x}, journal={Innovations in Systems and Software Engineering }, pages={517-531}, volume={21}}
+
+
