+++ docs/paper/paper.md
@@ -67,13 +67,13 @@ bibliography: paper.bib
 ---
 
 # Summary
-The environmental impact of research computing is increasingly a topic of concern for researchers. One of the main contributors of compute-related greenhouse gas (GHG) emissions is the production of electricity to power digital research infrastructures (DRIs). The carbon footprint per unit of electricity consumption (called carbon intensity) depends mostly on the grid's energy mix i.e. the share of renewable vs. high-carbon production methods, which varies greatly with time and location. Here, we describe the Climate Aware Task Scheduler (CATS) which lets researchers schedule computing when low-carbon electricity is available. CATS leverages carbon intensity forecasts, e.g. from the power grid operators, to find the best window of time to run a particular job with minimal climate cost. CATS also provides an assessment of the carbon savings due to delaying compute as opposed to running it immediately. We also demonstrate how the tool benefits researchers in the UK using the country's national grid's forecasts.
+The environmental impact of research computing is increasingly a topic of concern for researchers. One of the main contributors of compute-related greenhouse gas (GHG) emissions is the production of electricity to power digital research infrastructures (DRIs). The carbon footprint per unit of electricity consumption (called carbon intensity) depends mostly on the grid's energy mix i.e. the share of renewable vs. high-carbon production methods, which varies greatly with time and location. Here, we describe the Climate Aware Task Scheduler (CATS) which lets researchers schedule computing when low-carbon electricity is available. CATS leverages carbon intensity forecasts, e.g. from the power grid operators, to find the best window of time to run a particular job with minimal climate cost. CATS also provides an assessment of the carbon savings due to delaying compute as opposed to running it immediately. We demonstrate how the tool benefits researchers in the UK using forecasts of the National Grid's regional carbon intensity.
 
 # Statement of need
 
-The climate impact of research computing, computer science, and computational science is hard to understate. Computing impacts the environment in many ways, from water and abiotic resource usage to GHG emissions from energy consumption and electronic waste. For many researchers, the CO$_2$e (CO$_2$-equivalent, the usual metric of carbon footprint aggregating the main GHGs) cost associated with running their models is larger than that of any other aspects of their life [@PortegiesZwart_2020]. For example, the global carbon footprint of data centres is estimated at 126 Mt CO$_2$e [@Malmodin2024], equivalent to the entire American commercial aviation sector, and individual computing projects can reach dozens, if not hundreds, of tonnes of CO$_2$e [@PortegiesZwart_2020; @GrealeyLannelongue2022; @Luccioni_Viguier_Ligozat_2022]. This is becoming clear to researchers and funders, who are exploring different approaches to reducing and accounting for the climate impact of the research they commission [@Weber_2024; @Juckes23; @Lannelongue_Fropier_Matencio_2025]. There is a clear need to provide tools for motivated researchers to minimise the detrimental contribution of their research on the climate emergency, and to educate researchers around this impact.
[…]
