### 2025-06-24 9444a36
Update plot to use current implementation
Minor change in x axis label and fix the time
and place in the caption
