### 2025-06-24 1d03faf
Fix the [redacted] bibtex
