+++ docs/paper/paper.bib
@@ -85,8 +85,9 @@ series = {SC-W '23}
 
  @misc{MacBookPro, author={{Apple Inc}}, year={2024}, url={https://www.apple.com/environment/pdf/products/notebooks/MacBook_Pro_16-inch_PER_Oct2024.pdf}, language={en} }
 
- @misc{Norlander, author={Nordlander, T. F. and Lybecker, A.}, year={2023}, url={https://devblogs.microsoft.com/ise/saving-co2-using-location-and-time-shifting-in-azure/}, title={{Saving CO2 using location and time shifting in Azure
-}}
+ @misc{Norlander, author={Nordlander, T. F. and Lybecker, A.}, year={2023}, 
+       url={https://devblogs.microsoft.com/ise/saving-co2-using-location-and-time-shifting-in-azure/},
+       title={{Saving CO2 using location and time shifting in Azure}}}
 
  @article{Wiesner2021, author={Wiesner, P. and Behnke, I. and Scheinert, D. and Gontarska, K. and Thamsen, L.}, 
           year={2021}, title={{Let's wait awhile: how temporal workload shifting can reduce carbon emissions in the cloud}}, 
