### 2025-06-24 4c01a90
Update CONTRIBUTING to make sense for CATS
Also make link to CODE_OF_CONDUCT explicit
