### 2025-06-24 2ab7aad
Update contributing document in the docs
