+++ docs/source/contributing.rst
@@ -5,7 +5,109 @@
 Contributing
 ============
 
-We welcome contributions from the community!
+First off, thanks for taking the time to contribute!
 
-If you find a bug or have an idea for a new feature, please open an
-issue on `our GitHub repository <GitHubrepo_>`_ or submit a pull request.
+All types of contributions are encouraged and valued. See the below for different ways to help and details about how this project handles them. Please make sure to read the relevant section before making your contribution. It will make it a lot easier for us maintainers and smooth out the experience for all involved. The community looks forward to your contributions. All contributors are expected to abide by our code of conduct (see CODE_OF_CONDUCT.md in the repository). 
+
+.. NOTE::
+  And if you like the project, but just don't have time to contribute, that's fine. There are other easy ways to support the project and show your appreciation, which we would also be very happy about:
+  * Star the project
+  * Write about it
+  * Refer this project in your project's readme
+  * Mention the project at local meetups and tell your friends/colleagues
+
+
+I Have a Question
+-----------------
+
+.. NOTE::
+  If you want to ask a question, we assume that you have read the available documentation (at https://cats.readthedocs.io/).
+
+Before you ask a question, it is best to search for existing issues (see https://github.com/GreenScheduler/cats/issues) that might help you. In case you have found a suitable issue and still need clarification, you can write your question in this issue. It is also advisable to search the internet for answers first.
+
+If you then still feel the need to ask a question and need clarification, we recommend the following:
+
+* Open an issue (https://github.com/GreenScheduler/cats/issues/new).
+* Provide as much context as you can about what you're running into.
+* Provide project and platform versions (python version etc), depending on what seems relevant.
+* We use labels on GitHub to manage issues. Please add the "question" label to your issue.
+
+We will then take care of the issue as soon as possible. 
+
+
+I Want To Contribute
+--------------------
+
+.. NOTE::
+  When contributing to this project, you must agree that you have authored 100% of the content, that you have the necessary rights to the content and that the content you contribute may be provided under the project licence.
+
+Reporting Bugs
+--------------
+
+Before Submitting a Bug Report
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+A good bug report shouldn't leave others needing to chase you up for more information. Therefore, we ask you to investigate carefully, collect information and describe the issue in detail in your report. Please complete the following steps in advance to help us fix any potential bug as easily as possible.
+
+* Make sure that you are using the latest version.
+* Determine if your bug is really a bug and not an error on your side e.g. using incompatible environment components/versions (Make sure that you have read the [documentation](https://cats.readthedocs.io/). If you are looking for support, you might want to check [this section](#i-have-a-question)).
+* To see if other users have experienced (and potentially already solved) the same issue you are having, check if there is not already a bug report existing for your bug or error in the [bug tracker](https://github.com/GreenScheduler/cats/issues?q=label%3Abug).
+* Also make sure to search the internet (including Stack Overflow) to see if users outside of the GitHub community have discussed the issue.
+* Collect information about the bug:
+
+  * OS, Platform and Version (Windows, Linux, macOS, x86, ARM)
+  * Version of python and your python environment, package manager, and how you installed CATS.
+  * Possibly your input and the output
+  * Can you reliably reproduce the issue? And can you also reproduce it with older versions?
+
+
+How Do I Submit a Good Bug Report?
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+.. NOTE::
[…]
