### 2025-06-24 049de07
Minor fixes to contributing
