### 2025-06-24 cc0c2f0
More fixes
