+++ docs/source/contributing.rst
@@ -48,8 +48,8 @@ Before Submitting a Bug Report
 A good bug report shouldn't leave others needing to chase you up for more information. Therefore, we ask you to investigate carefully, collect information and describe the issue in detail in your report. Please complete the following steps in advance to help us fix any potential bug as easily as possible.
 
 * Make sure that you are using the latest version.
-* Determine if your bug is really a bug and not an error on your side e.g. using incompatible environment components/versions (Make sure that you have read the [documentation](https://cats.readthedocs.io/). If you are looking for support, you might want to check [this section](#i-have-a-question)).
-* To see if other users have experienced (and potentially already solved) the same issue you are having, check if there is not already a bug report existing for your bug or error in the [bug tracker](https://github.com/GreenScheduler/cats/issues?q=label%3Abug).
+* Determine if your bug is really a bug and not an error on your side e.g. using incompatible environment components/versions. Make sure that you have read the documentation at https://cats.readthedocs.io/). If you are looking for support, you might want to check the section above.
+* To see if other users have experienced (and potentially already solved) the same issue you are having, check if there is not already a bug report existing for your bug or error in the bug tracker (see https://github.com/GreenScheduler/cats/issues?q=label%3Abug).
 * Also make sure to search the internet (including Stack Overflow) to see if users outside of the GitHub community have discussed the issue.
 * Collect information about the bug:
 
