### 2025-06-25 b5dbce5
Add info on running tests etc.
