+++ docs/source/contributing.rst
@@ -155,7 +155,7 @@ should be discussed ahead of time (via a PR changing the version string, see 1 b
 
  1. Merge a pull request onto main that updates the CATS version number ``__version__`` in ``__init__.py`` and adds any release notes / key changes to the documentation. We use a "major.minor.patch" semantic versioning scheme; for bug fixes etc. bump the patch number, for significant new features bump the minor version number, for changes that break previous behavior update the major version number.
  2. Check that all tests have passed after the merge and that the "latest" documentation at read the docs is updated.
- 3. Create a release via the GitHub web interface. This involves creating a new tag ("v1.2.3" for version "1.2.3"), giving the release a name (just "1.2.3"), and adding short release notes using markdown as needed. make sure this is "set as the latest release".
+ 3. Create a release via the GitHub web interface. This involves creating a new tag ("v1.2.3" for version "1.2.3"), giving the release a name (just "1.2.3"), and adding short release notes using markdown as needed. Make sure this is "set as the latest release".
  4. After a short time you should be able to check that the new release exists on PyPi and is documented in the stable docs on read the docs.   
 
 Attribution
