### 2025-06-25 b48c481
PyPi should be PyPI
Co-authored-by: Sadie L. Bartholomew <sadie.bartholomew@ncas.ac.uk>
