### 2025-06-25 03db9a7
PyPi should be PyPI
Co-authored-by: Sadie L. Bartholomew <sadie.bartholomew@ncas.ac.uk>
