### 2025-06-25 #8251 @LiamPattinson
Thanks @andreww, I've had a look over the latest changes and I'll be happy to approve once the `--plot` functionality is merged.
