### 2025-06-26 #8251 @andreww
As far as I can see everything is now done - I've pushed out a new version to PyPI which includes `--plot`, information about contributing, and we've got some words and references about other timeshifting approaches in the paper. Do shout (or reopen issues) if I've missed anything!
