### 2025-06-27 3f85ac7
Turn the Turing way into a reference, fix doc link
During review we moved our documentation onto readthedocs, so
update that link. We added a reference to The Turing Way, but
didn't include the proper citation. Both are fixed. We had
two links to documentation. Remove one to avoid duplication.
