+++ docs/paper/paper.md
@@ -73,11 +73,11 @@ The environmental impact of research computing is increasingly a topic of concer
 
 The climate impact of research computing, computer science, and computational science is hard to understate. Computing impacts the environment in many ways, from water and abiotic resource usage to GHG emissions from energy consumption and electronic waste. For many researchers, the CO$_2$e (CO$_2$-equivalent, the usual metric of carbon footprint aggregating the main GHGs) cost associated with running their models is larger than that of any other aspects of their life [@PortegiesZwart_2020]. The global carbon footprint of data centres is estimated at 126 Mt CO$_2$e [@Malmodin2024], equivalent to the entire American commercial aviation sector, and individual computing projects can reach dozens, if not hundreds, of tonnes of CO$_2$e [@PortegiesZwart_2020; @GrealeyLannelongue2022; @Luccioni_Viguier_Ligozat_2022]. This is becoming clear to researchers and funders, who are exploring different approaches to reducing and accounting for the climate impact of the research they commission [@Weber_2024; @Juckes23; @Lannelongue_Fropier_Matencio_2025]. There is a clear need to provide tools for motivated researchers to minimise the detrimental contribution of their research on the climate emergency, and to educate researchers around this impact.
 
-There are several approaches that have been proposed and/or implemented to reduce the carbon footprint of computing, which are summarised in The Turing Way (https://book.the-turing-way.org/ethical-research/activism/activism-env-impact) and in the CATS documentation (https://greenscheduler.github.io/cats/). One of these is time shifting the computation such that it runs when the power supplied by the local electricity grid is dominated by renewable generation such as on windy or sunny days rather than periods where legacy fossil fuel generation dominates. The potential impact of minimally-invasive time-shifting approaches has been shown to result in significantly reduced carbon footprints [27% in one AI benchmark @Dodge2022]. CATS helps researchers timeshift their own computation such that it is scheduled when the forecast carbon intensity of the power grid is minimised. This timeshifting approach has been applied in the cloud [@Wiesner2021], for large-language model training [@Jagannadharao2025], and to take account of carbon intensity when scaling Kubernetes workloads on Azure [@Norlander], but we are not aware of a similar user facing tool. 
+There are several approaches that have been proposed and/or implemented to reduce the carbon footprint of computing, which are summarised in The Turing Way [@TuringWay] and in the CATS documentation. One of these is time shifting the computation such that it runs when the power supplied by the local electricity grid is dominated by renewable generation such as on windy or sunny days rather than periods where legacy fossil fuel generation dominates. The potential impact of minimally-invasive time-shifting approaches has been shown to result in significantly reduced carbon footprints [27% in one AI benchmark @Dodge2022]. CATS helps researchers timeshift their own computation such that it is scheduled when the forecast carbon intensity of the power grid is minimised. This timeshifting approach has been applied in the cloud [@Wiesner2021], for large-language model training [@Jagannadharao2025], and to take account of carbon intensity when scaling Kubernetes workloads on Azure [@Norlander], but we are not aware of a similar user facing tool. 
 
 # Approach and functionality
 
[…]
