+++ docs/paper/paper.bib
@@ -98,4 +98,5 @@ series = {SC-W '23}
           year={2025}, title={Timeshifting strategies for carbon-efficient long-running large language model training},
           doi={10.1007/s11334-023-00546-x}, journal={Innovations in Systems and Software Engineering }, pages={517-531}, volume={21}}
 
-
+@book{TuringWay, author={{The Turing Way Community}}, title={{The Turing Way: A handbook for reproducible, ethical and collaborative research (1.2.3)}},
+      year={2025}, doi={10.5281/zenodo.15213042}}
