### 2025-06-27 #8251 @andreww
@jatkinson1000 I've upload version 1.1.0 (tagged "v1.1.0" in the repository) to Zenodo with DOI for this version of: [10.5281/zenodo.15753041](https://doi.org/10.5281/zenodo.15753041).
