### 2025-07-07 #8251 @jatkinson1000
@andreww It looks like there is an invalid DOI in the paper (Wiesner et al (2021))
Please could you fix this and let me know when this is done.

Please also check the articles without DOIs that the JOSS bot thinks should have them and add if available.
