+++ docs/paper/paper.bib
@@ -91,7 +91,7 @@ series = {SC-W '23}
 
  @article{Wiesner2021, author={Wiesner, P. and Behnke, I. and Scheinert, D. and Gontarska, K. and Thamsen, L.}, 
           year={2021}, title={{Let's wait awhile: how temporal workload shifting can reduce carbon emissions in the cloud}}, 
-          doi={10.1145/3464298.349339}, journal={{Middleware '21: Proceedings of the 22nd International Middleware Conference}}, 
+          doi={10.1145/3464298.3493399}, journal={{Middleware '21: Proceedings of the 22nd International Middleware Conference}}, 
           pages={{260-272}}}
 
  @article{Jagannadharao2025, author={Jagannadharao, A. and Beckage, N. and Nafus, D. and Chamberlin, S.},
