+++ docs/paper/paper.bib
@@ -64,7 +64,8 @@ series = {SC-W '23}
 
  @article{Malmodin2024, title={ICT sector electricity consumption and greenhouse gas emissions – 2020 outcome}, ISSN={0308-5961}, DOI={10.1016/j.telpol.2023.102701}, abstractNote={The Information and Communication Technology (ICT) sector has gained much attention in the discussions on climate change, as it could impact global emissions both positively and negatively. The objective of the present study is to provide estimates for the 2020 use stage electricity consumption and ICT sector’s total lifecycle greenhouse gas (GHG) emissions divided in three main parts: user devices including internet-of-things, networks and data centers. The study builds on a high number of data sources including measured and reported data from 150 companies that is estimated to cover about 80% of network subscriptions, about 55% of data center electricity, and about 35% of upstream GHG emissions. To understand the development, the results are put into the perspective of earlier studies. In conclusion, the ICT sector used about 4% of the global electricity in the use stage and represented about 1.4% of the global GHG emissions in 2020. The use stage electricity consumption and the total GHG emissions have increased since 2015, but the impact per subscription has decreased. The user devices accounted for over half of all GHG emissions, with equal parts relating to use stage and other lifecycle stages. For networks and data centers, the use stage GHG emissions are dominating. The electricity consumption and GHG emissions are also estimated for the closely related areas Entertainment and Media (including e.g., TVs), paper media, and cryptocurrencies.}, journal={Telecommunications Policy}, author={Malmodin, Jens and Lövehagen, Nina and Bergmark, Pernilla and Lundén, Dag}, year={2024}, month=jan, pages={102701} }
 
- @article{Luccioni_Viguier_Ligozat_2022, title={Estimating the Carbon Footprint of BLOOM, a 176B Parameter Language Model}, url={http://arxiv.org/abs/2211.02001}, abstractNote={Progress in machine learning (ML) comes with a cost to the environment, given that training ML models requires significant computational resources, energy and materials. In the present article, we aim to quantify the carbon footprint of BLOOM, a 176-billion parameter language model, across its life cycle. We estimate that BLOOM’s final training emitted approximately 24.7 tonnes of~carboneq~if we consider only the dynamic power consumption, and 50.5 tonnes if we account for all processes ranging from equipment manufacturing to energy-based operational consumption. We also study the energy requirements and carbon emissions of its deployment for inference via an API endpoint receiving user queries in real-time. We conclude with a discussion regarding the difficulty of precisely estimating the carbon footprint of ML models and future research directions that can contribute towards improving carbon emissions reporting.}, note={arXiv:2211.02001 [cs]}, number={arXiv:2211.02001}, publisher={arXiv}, author={Luccioni, Alexandra Sasha and Viguier, Sylvain and Ligozat, Anne-Laure}, year={2022}, month=nov }
+ @article{Luccioni_Viguier_Ligozat_2022, title={Estimating the Carbon Footprint of BLOOM, a 176B Parameter Language Model}, 
[…]
