+++ docs/paper/paper.bib
@@ -1,4 +1,4 @@
-@article{Bruce21a,
+article{Bruce21a,
        url = {https://github.com/carbon-intensity/methodology/raw/master/Carbon%20Intensity%20Forecast%20Methodology.pdf},
        Author = {{Bruce}, A.~R.~W. and {Ruff}, L. and {Kelloway}, J. and {MacMillan}, F. and {Rogers}, A.},
        Title = {Carbon intensity forecast methodology},
@@ -7,7 +7,7 @@
 }
 
 
-@article{Bruce21b,
+article{Bruce21b,
        url = {https://github.com/carbon-intensity/methodology/raw/master/Regional%20Carbon%20Intensity%20Forecast%20Methodology.pdf},
        Author = {{Bruce}, A.~R.~W. and {Ruff}, L. and {Kelloway}, J. and {MacMillan}, F. and {Rogers}, A.},
        Title = {Carbon intensity methodology regional carbon intensity},
@@ -72,15 +72,21 @@ series = {SC-W '23}
  @article{GrealeyLannelongue2022, title={The Carbon Footprint of Bioinformatics}, rights={All rights reserved}, ISSN={0737-4038, 1537-1719}, DOI={10.1093/molbev/msac034}, abstractNote={ABSTRACT
             Bioinformatic research relies on large-scale computational infrastructures which have a non-zero carbon footprint but so far, no study has quantified the environmental costs of bioinformatic tools and commonly run analyses. In this work, we estimate the carbon footprint of bioinformatics (in kilograms of CO2 equivalent units, kgCO2e) using the freely available Green Algorithms calculator (www.green-algorithms.org). We assessed (i) bioinformatic approaches in genome-wide association studies (GWAS), RNA sequencing, genome assembly, metagenomics, phylogenetics and molecular simulations, as well as (ii) computation strategies, such as parallelisation, CPU (central processing unit) vs GPU (graphics processing unit), cloud vs. local computing infrastructure and geography. In particular, we found that biobank-scale GWAS emitted substantial kgCO2e and simple software upgrades could make it greener, e.g. upgrading from BOLT-LMM v1 to v2.3 reduced carbon footprint by 73%. Moreover, switching from the average data centre to a more efficient one can reduce carbon footprint by ∼34%. Memory over-allocation can also be a substantial contributor to an algorithm’s greenhouse gas (GHG) emissions. The use of faster processors or greater parallelisation reduces running time but can lead to greater carbon footprint. Finally, we provide guidance on how researchers can reduce power consumption and minimise kgCO2e. Overall, this work elucidates the carbon footprint of common analyses in bioinformatics and provides solutions which empower a move toward greener research.}, journal={Molecular Biology and Evolution}, author={Grealey, Jason and Lannelongue, Loïc and Saw, Woei-Yuh and Marten, Jonathan and Méric, Guillaume and Ruiz-Carmona, Sergio and Inouye, Michael}, year={2022}, month=feb, pages={msac034}, language={en} }
 
[…]
