+++ docs/paper/paper.bib
@@ -90,7 +90,7 @@ abstractNote={Artificial intelligence (AI) is in the media spotlight for its pot
 
  @book{Lannelongue_Vegad_Dorn_2024, title={Environmental Impacts of the UK Digital Sector - Data Driven Report}, url={https://zenodo.org/records/14831758}, DOI={10.17863/CAM.114743}, abstractNote={The digital sector has been and continues to be a catalyst for new capabilities and efficiencies, transforming the way society operates and interacts. The environmental impacts of this sector, such as greenhouse gas (GHG) emissions and resources depletion, are significant and require careful consideration. The first essential step to inform future policy making is to quantify these impacts through a detailed inventory for the UK. This is something that had not been conducted before. Commissioned by the Department for Culture, Media and Sport (DCMS) and the Department for Science, Innovation and Technology (DSIT), this report provides a robust analysis of the electricity consumption, GHG emissions and wider environmental impacts associated with the UK’s digital sector. It also investigates future trends through case studies and scenario projections. The research was conducted in partnership with Frazer Nash Consultancy, the University of Cambridge, and the University of Bristol.}, author={Lannelongue, Loïc and Vegad, Vinesh and Dorn, Emily}, year={2024}, month=may, language={eng} }
 
- @misc{MacBookPro, author={{Apple Inc}}, year={2024}, url={https://www.apple.com/environment/pdf/products/notebooks/MacBook_Pro_16-inch_PER_Oct2024.pdf}, language={en} }
+ @misc{MacBookPro, title={Product Environmental Report MacBook Pro 16-inch}, author={{Apple Inc}}, year={2024}, url={https://www.apple.com/environment/pdf/products/notebooks/MacBook_Pro_16-inch_PER_Oct2024.pdf}, language={en} }
 
  @misc{Norlander, author={Nordlander, T. F. and Lybecker, A.}, year={2023}, 
        url={https://devblogs.microsoft.com/ise/saving-co2-using-location-and-time-shifting-in-azure/},
