### 2025-07-07 #8251 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
