### 2025-07-08 a406c62
Remove address details
JOSS style does not need full address details
