+++ docs/paper/paper.md
@@ -35,33 +35,33 @@ authors:
     orcid: 0009-0008-3463-2694
     affiliation: 13
 affiliations:
- - name: National Centre for Atmospheric Science, Fairbairn House, 71-75 Clarendon Road, Leeds, LS2 9PH, United Kingdom
+ - name: National Centre for Atmospheric Science, United Kingdom
    index: 1
- - name: Department of Meteorology, Brian Hoskins Building, University of Reading, Whiteknights campus, RG6 6ET, United Kingdom
+ - name: Department of Meteorology, University of Reading, United Kingdom
    index: 2
- - name: School of Psychology, University of Sussex, Brighton, United Kingdom
+ - name: School of Psychology, University of Sussex, United Kingdom
    index: 3
- - name: Oxford Research Software Engineering Group, Doctoral Training Centre, University of Oxford, 1-4 Keble Road, Oxford, OX1 3NP, United Kingdom
+ - name: Oxford Research Software Engineering Group, University of Oxford, United Kingdom
    index: 4
- - name: Bayesic Research, Ithaca, NY, USA
+ - name: Bayesic Research, Ithaca, USA
    index: 5
- - name: British Heart Foundation Cardiovascular Epidemiology Unit, Department of Public Health and Primary Care, University of Cambridge, Cambridge, United Kingdom
+ - name: British Heart Foundation Cardiovascular Epidemiology Unit, University of Cambridge, United Kingdom
    index: 6
- - name: Victor Phillip Dahdaleh Heart and Lung Research Institute, University of Cambridge, Cambridge, United Kingdom
+ - name: Victor Phillip Dahdaleh Heart and Lung Research Institute, University of Cambridge, United Kingdom
    index: 7
- - name: Health Data Research UK Cambridge, Wellcome Genome Campus and University of Cambridge, Cambridge, United Kingdom
+ - name: Health Data Research UK Cambridge, United Kingdom
    index: 8
- - name: CNRM, Université de Toulouse, Météo-France, CNRS, Toulouse, France
+ - name: CNRM, Université de Toulouse, France
    index: 9
- - name: Netherlands eScience Center, Science Park 402 (Matrix THREE), 1098 XH Amsterdam, Netherlands
+ - name: Netherlands eScience Center, Netherlands
    index: 10
- - name: National Oceanography Centre, Joseph Proudman Building, 6 Brownlow Street, Liverpool, L3 5DA, United Kingdom
+ - name: National Oceanography Centre, Liverpool, United Kingdom
    index: 11
- - name: Department of Earth Sciences, University of Oxford, South Parks Road, Oxford, OX1 3AN, United Kingdom
+ - name: Department of Earth Sciences, University of Oxford, United Kingdom
    index: 12
- - name: National Oceanography Centre, European Way, Southampton, SO14 3ZH, United Kingdom
+ - name: National Oceanography Centre, Southampton, United Kingdom
    index: 13
-date: 14 April 2025
+date: 8 July 2025
 bibliography: paper.bib
 
 ---
