### 2025-07-08 #8251 @danielskatz
Congratulations to @andreww (Andrew Walker) and co-authors on your JOSS publication!!

JOSS depends on volunteers, and we couldn't do this without you
